Added a DFSBackend extension point
A new INI option, eZDFSClusteringSettings.DFSBackend is introduced,
that defaults to eZDFSFileHandlerDFSBackend, the historical handler.

It can be set to another class that implements eZDFSFileHandlerDFSBackendInterface.

If this class implements eZDFSFileHandlerFactoryDFSBackendInterface, the instance
will be created by calling the static factory() method. If it doesn't, it will
be constructed normally. In both cases, no arguments are passed.